pokepoint/requirements.txt
# pip install -r requirements.txt
django==3.2
djangorestframework==3.12.4
psycopg2==2.8.6
django-crispy-forms==1.12.0


